requirements.txt
ipython==7.32.0
matplotlib==3.5.0
mplcursors==0.5.1
numpy==1.20.3
pandas==1.3.5
pickle5==0.0.12
Pillow==9.0.1
scikit_learn==1.0.2
ttkbootstrap==0.5.1
ttkwidgets==0.12.1
